+++ core/src/com/bot4s/zmatrix/core/RequestAuth.scala
@@ -1,0 +1,7 @@
+package com.bot4s.zmatrix.core
+
+sealed trait RequestAuth
+object RequestAuth {
+  case object NoAuth                  extends RequestAuth
+  case class TokenAuth(token: String) extends RequestAuth
+}
